refactor: Make TastingListHeader generic
