+++ src/containers/organisms/TastingListHeader.tsx
@@ -1,6 +1,7 @@
 import React, { FC } from 'react';
 import { useHistory } from 'react-router-dom';
-import TastingSheetHeader from 'components/organisms/TastingListHeader';
+import { Dropdown } from 'semantic-ui-react';
+import TastingSheetHeader from 'components/organisms/PageHeader';
 
 const EnhancedTastingListHeader: FC = () => {
   const history = useHistory();
@@ -9,7 +10,28 @@ const EnhancedTastingListHeader: FC = () => {
     history.push('/tastingSheet');
   };
 
-  return <TastingSheetHeader handleClick={handleClick} />;
+  const actionElement = (
+    <Dropdown text="テイスティングシート作成">
+      <Dropdown.Menu>
+        <Dropdown.Item
+          text="赤ワイン"
+          label={{ color: 'purple', empty: true, circular: true }}
+          onClick={handleClick}
+        />
+        <Dropdown.Item
+          text="白ワイン"
+          label={{ color: 'yellow', empty: true, circular: true }}
+        />
+      </Dropdown.Menu>
+    </Dropdown>
+  );
+
+  return (
+    <TastingSheetHeader
+      headerTitle="テイスティングリスト"
+      actionElement={actionElement}
+    />
+  );
 };
 
 export default EnhancedTastingListHeader;
